requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra=dev --no-emit-index-url --output-file=requirements.txt --strip-extras
#
attrs==25.1.0
    # via flake8-bugbear
black==25.1.0
    # via bnovate (setup.py)
blinker==1.9.0
    # via flask
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   black
    #   flask
    #   pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
flake8==7.1.2
    # via
    #   bnovate (setup.py)
    #   flake8-bugbear
    #   flake8-docstrings
    #   pep8-naming
flake8-bugbear==24.12.12
    # via bnovate (setup.py)
flake8-docstrings==1.7.0
    # via bnovate (setup.py)
flask==3.1.0
    # via bnovate (setup.py)
idna==3.10
    # via requests
isort==6.0.1
    # via bnovate (setup.py)
itsdangerous==2.2.0
    # via flask
jinja2==3.1.5
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
mypy-extensions==1.0.0
    # via black
numpy==2.2.3
    # via
    #   bnovate (setup.py)
    #   pandas
    #   shapely
packaging==24.2
    # via
    #   black
    #   build
pandas==2.2.3
    # via bnovate (setup.py)
pathspec==0.12.1
    # via black
pep8-naming==0.14.1
    # via bnovate (setup.py)
pip-tools==7.4.1
    # via bnovate (setup.py)
platformdirs==4.3.6
    # via black
psycopg2-binary==2.9.10
    # via bnovate (setup.py)
pycodestyle==2.12.1
    # via flake8
pydocstyle==6.3.0
    # via flake8-docstrings
pyflakes==3.2.0
    # via flake8
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pandas
pytz==2025.1
    # via pandas
pyyaml==6.0.2
    # via bnovate (setup.py)
requests==2.32.3
    # via bnovate (setup.py)
shapely==2.0.7
    # via bnovate (setup.py)
six==1.17.0
    # via python-dateutil
snowballstemmer==2.2.0
    # via pydocstyle
tomli==2.2.1
    # via
    #   black
    #   build
    #   pip-tools
typing-extensions==4.12.2
    # via black
tzdata==2025.1
    # via pandas
urllib3==2.3.0
    # via requests
werkzeug==3.1.3
    # via flask
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
